Fix missing python3.7 on Ubuntu 20.04 LTS
When doing a clean run on Ubuntu 20.04 LTS the maubot role attempts to
install `python3.7` and the corresponding venv package. In Ubuntu 20.04
however, `python3.7` is no longer available. The default is `python3.8`.
Therefore I would propose switching over to `python3` in general.

This has been testet to work on Ubuntu 20.04 LTS, but not on older
versions. Therefore, further tests need to be done.