Initial prometheus alertmanager setup
The two hosts will monitor each other using prometheus. There is only a
short retention period and no grafana setup since this will A save on
storage / resources (of which our supply is not unlimited) and B I don't
see any value (for us) we could get out of long term metrics.

I optet not to export special Matrix or Keycloak metrics, mainly due to
simplicity. The health of these services is simply being monitored by
looking at the share of http response codes. Together with a check for
our hosts being up, the disks not being full and no systemd services
crashing this should cover all of the outages we've had so far.

Close #9