Lower high memory usage bar
Earlier today (and over the last couple of days) the memory usage on
hopper (matrix host) rose to the level that users were unable to sign
in. However the monitoring has not caught that. Settings a bar of 75%
free would have catched that, but (regarding the last 1w) no accidental
hits could be observed.